KAN-39 feat: move sign-up validation schema to shared/schema
