--- src/shared/lib/zod/index.ts
@@ -1,1 +1,0 @@
-export * from './authSchemas'
